pgraster: refactor gt-imagemosaic-jdbc package and remove import tool
Rename package org.geotools.gce.imagemosaic.jdbc as
org.geotools.gce.pgraster.

Remove ability to import tiles from disk while creating
the configuration objects. By one side, it's to be done
by external tooling, and by the other, may it be performed
by GeoServer, it should be a separate, explicit tool, rather
than an automagic behavior while creating the config.
