+++ src/community/pgraster/pom.xml
@@ -16,7 +16,7 @@
   <groupId>org.geoserver.community</groupId>
   <artifactId>gs-pgraster</artifactId>
   <packaging>jar</packaging>
-  <name>PGRaster DataStore Extension for automatic configuration</name>
+  <name>PGRaster CoverageStore Extension</name>
 
   <dependencies>
     <dependency>
@@ -26,17 +26,14 @@
     </dependency>
     <dependency>
       <groupId>org.geotools</groupId>
-      <artifactId>gt-epsg-hsql</artifactId>
+      <artifactId>gt-jdbc</artifactId>
       <version>${gt.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.ant</groupId>
-      <artifactId>ant</artifactId>
     </dependency>
     <dependency>
       <groupId>org.geotools</groupId>
-      <artifactId>gt-jdbc</artifactId>
+      <artifactId>gt-epsg-hsql</artifactId>
       <version>${gt.version}</version>
+      <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>commons-dbcp</groupId>
@@ -61,10 +58,6 @@
       <version>${project.version}</version>
       <classifier>tests</classifier>
       <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>com.ibm.icu</groupId>
-      <artifactId>icu4j</artifactId>
     </dependency>
   </dependencies>
 </project>
