+++ src/community/pgraster/src/main/java/org/geotools/gce/pgraster/config/SpatialExtension.java
@@ -11,7 +11,7 @@
  * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * Lesser General Public License for more details.
  */
-package org.geotools.gce.imagemosaic.jdbc;
+package org.geotools.gce.pgraster.config;
 
 /**
  * Enum for support spatial extension
