+++ src/community/pgraster/src/main/java/org/geotools/gce/pgraster/reader/ImageLevelInfo.java
@@ -14,7 +14,7 @@
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  *    Lesser General Public License for more details.
  */
-package org.geotools.gce.imagemosaic.jdbc;
+package org.geotools.gce.pgraster.reader;
 
 import java.io.ByteArrayOutputStream;
 import java.io.PrintWriter;
