+++ src/community/pgraster/src/main/resources/META-INF/services/org.geotools.coverage.grid.io.GridFormatFactorySpi
@@ -1,1 +1,1 @@
-org.geotools.gce.imagemosaic.jdbc.ImageMosaicJDBCFormatFactory
+org.geotools.gce.pgraster.PostgisRasterFormatFactory
