+++ src/community/pgraster/src/main/resources/org/geotools/gce/imagemosaic/pgraster/coverage.pgraster.template.xml
